### 2020-07-09 #2455 @whedon
[PRE REVIEW]: bridge-sim: A Python library for concrete slab bridge simulation
**Submitting author:** @barischrooneyj (<a href="http://orcid.org/0000-0002-8441-212X">Jeremy Barisch-Rooney</a>)
**Repository:** <a href="https://github.com/barischrooneyj/bridge-sim" target ="_blank">https://github.com/barischrooneyj/bridge-sim</a>
**Version:** v0.0.7
