### 2020-07-09 #2455 @kyleniemeyer
Hi @barischrooneyj, thanks for your submission to JOSS. While we are getting started, could you add a Statement of Need to your article? This is now required for articles, as described in our [documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Thanks!
