### 2020-07-09 #2455 @poulson
I'm weary about being the editor on this and would prefer that a civil engineer do so. I would be comfortable reviewing this if it was over the FEM itself, but I feel there are deeper constraints here due to the safety implications of bridge design and my blindness to which types of models are considered in bounds.
