### 2020-07-15 #2455 @skreimeyer
Patrick,

I will look over the code, but it may be this weekend before I get to it.

Regards,
Sam

On Mon, Jul 13, 2020, 9:33 AM Patrick Diehl <notifications@github.com>
wrote:

> Hey, @TheBeachLab <https://github.com/TheBeachLab> @skreimeyer
> <https://github.com/skreimeyer> @dukenmarga
> <https://github.com/dukenmarga> @ebrahimraeyat
> <https://github.com/ebrahimraeyat> @manparvesh
> <https://github.com/manparvesh> @ipsumium <https://github.com/ipsumium>
> would you be interested in reviewing this paper?
>
