### 2020-07-20 #2496 @diehlpk
@barischrooneyj

Following things:

1. Can you extend 1 TNO with the full name? Most people might not know what TNO means.
2. State of the field: Do the authors describe how this software compares to other commonly-used packages? I feel that this point is missing in the paper. Can you add more details there?

I leave the remaining check boxes to the reviewers :)
