### 2020-07-27 #2496 @barischrooneyj
> @barischrooneyj
>
> Following things:
>
> 1. Can you extend 1 TNO with the full name? Most people might not know what TNO means.
> 2. State of the field: Do the authors describe how this software compares to other commonly-used packages? I feel that this point is missing in the paper. Can you add more details there?
> 3. Add countries (full name no abbreviation) to the affiliations
>
> I leave the remaining check boxes to the reviewers :)

@diehlpk @andybond13 I'm in the process of moving house/country, so I can get around to addressing your comments and I will also do another pass of the documentation, next week. I will let you know when that pass is done. Best, Jeremy
