### 2020-07-31 3377839
more informative error messages when OS cannot be run (particularly for Windows machines)
