+++ docker/tzdata.sh
@@ -1,7 +1,0 @@
-# set noninteractive installation
-export DEBIAN_FRONTEND=noninteractive
-# install tzdata package
-apt-get install -y tzdata
-# set your timezone
-ln -fs /usr/share/zoneinfo/America/New_York /etc/localtime
-dpkg-reconfigure --frontend noninteractive tzdata
