+++ scripts/docs.sh
@@ -1,5 +1,5 @@
 rm -rf docs/
-pipenv run pdoc --html --output-dir build bridge_sim
+poetry run pdoc --html --output-dir build bridge_sim
 mv build/bridge_sim docs/
 rm -rf build/
-pipenv run python scripts/docs-update.py
+poetry run python scripts/docs-update.py
