+++ scripts/mypy.sh
@@ -1,3 +1,3 @@
-pipenv run stubgen src/
+poetry run stubgen src/
 export MYPYPATH=~/cs/bridge-sim/out/src/
-pipenv run mypy src/bridge_sim/
+poetry run mypy src/bridge_sim/
