### 2020-08-07 dd5b093
FileNotFoundError if OpenSees not found
