+++ src/bridge_sim/configs/__init__.py
@@ -5,25 +5,31 @@ from typing import Callable, Optional
 
 from bridge_sim.bridges import bridge_705
 from bridge_sim.model import Config, Bridge
-from bridge_sim.sim.run.opensees import os_runner
+from bridge_sim.sim.run.opensees import OSRunner
 from bridge_sim.util import project_dir
 
 
 def opensees_default(
-    bridge: Callable[[], Bridge], os_exe: Optional[str] = None, **kwargs
+    bridge: Callable[[], Bridge],
+    exe_path: Optional[str] = None,
+    allow_no_exe: bool = False,
+    **kwargs,
 ) -> Config:
     """A Config using OpenSees for a given Bridge.
 
     Args:
         bridge: function to return a new Bridge.
-        os_exe: absolute path to OpenSees binary. Optional, if not given this
+        exe_path: absolute path to OpenSees binary. Optional, if not given this
             will look for OpenSees on the $PATH.
+        allow_no_exe: only useful for testing in environments where OpenSees is
+            not available, allows for the construction of a Config without
+            having OpenSees installed.
         kwargs: keyword arguments passed to the Config constructor.
 
     """
     return Config(
         bridge=bridge,
-        sim_runner=os_runner(os_exe),
+        sim_runner=OSRunner(exe_path, allow_no_exe=allow_no_exe),
         vehicle_data_path=os.path.join(project_dir(), "data/traffic/traffic.csv"),
         vehicle_pdf=[
             (2.4, 5),
@@ -41,11 +47,11 @@ def opensees_default(
 
 def test_config(msl: float = 10.0):
     """A Config used internally for testing."""
-    config = opensees_default(bridge_705(msl))
+    config = opensees_default(bridge_705(msl), allow_no_exe=True)
     config._root_generated_data_dir = os.path.join(
         "test-data/", config._root_generated_data_dir
     )
-    exe_found = config.sim_runner is not None
+    exe_found = config.sim_runner.exe_path is not None
     return config, exe_found
 
 
