### 2020-08-07 6160577
FileNotFoundError if OpenSees not found
