+++ scripts/format.sh
@@ -1,0 +1,1 @@
+bash -c 'shopt -s globstar && poetry run black -t py37 src/**/*.py'
