### 2020-08-07 2fd205c
Remove unused function
