### 2020-08-07 657a436
Move PermissionError to non-FE specific code
