+++ example.py
@@ -1,4 +1,6 @@
-# Responses to a point-load.
+########################################
+# Example 1: responses to a point-load #
+########################################
 
 import matplotlib.pyplot as plt
 from bridge_sim import bridges, configs, model, plot, sim
@@ -11,217 +13,233 @@ plot.top_view_bridge(config.bridge, piers=True)
 plt.tight_layout()
 plt.show()
 
-# Responses to a vehicle.
+############################################
+# Example 2: responses to a static vehicle #
+############################################
 
-# import matplotlib.pyplot as plt
-# from bridge_sim import bridges, configs, model, plot, sim
+import matplotlib.pyplot as plt
+from bridge_sim import bridges, configs, model, plot, sim
 
-# config = configs.opensees_default(bridges.bridge_narrow, shorten_paths=True)
-# point_loads = model.Vehicle(
-#     # Load intensity of each axle.
-#     load=[5000, 4000, 4000, 5000, 7000],
-#     # Distance between each pair of axles.
-#     axle_distances=[2, 2, 2, 1],
-#     # Width of each axle, distance between point loads.
-#     axle_width=2.5,
-#     # Speed of the vehicles.
-#     kmph=20,
-# ).point_load_pw(config=config, time=3.5, list=True)
-# responses = sim.responses.load(config, model.RT.YTrans, point_loads)
-# plot.contour_responses(config, responses, point_loads)
-# plot.top_view_bridge(config.bridge, piers=True)
-# plt.tight_layout()
-# plt.show()
+config = configs.opensees_default(bridges.bridge_narrow, shorten_paths=True)
+point_loads = model.Vehicle(
+    # Load intensity of each axle.
+    load=[5000, 4000, 4000, 5000, 7000],
+    # Distance between each pair of axles.
+    axle_distances=[2, 2, 2, 1],
+    # Width of each axle, distance between point loads.
+    axle_width=2.5,
+    # Speed of the vehicles.
+    kmph=20,
+).point_load_pw(config=config, time=3.5, list=True)
+responses = sim.responses.load(config, model.RT.YTrans, point_loads)
+plot.contour_responses(config, responses, point_loads)
+plot.top_view_bridge(config.bridge, piers=True)
+plt.tight_layout()
+plt.show()
 
-# Responses to pier settlement.
+###########################################
+# Example 3: responses to pier settlement #
+###########################################
 
-# import matplotlib.pyplot as plt
-# from bridge_sim import bridges, configs, sim, model, plot
+import matplotlib.pyplot as plt
+from bridge_sim import bridges, configs, sim, model, plot
 
-# config = configs.opensees_default(bridges.bridge_wide)
-# responses = sim.responses.load(
-#     config,
-#     model.RT.YTrans,
-#     pier_settlement=[model.PierSettlement(0, 1.2)]
-# )
-# plot.contour_responses(config, responses)
-# plot.top_view_bridge(config.bridge, piers=True)
-# plt.tight_layout()
-# plt.show()
+config = configs.opensees_default(bridges.bridge_wide)
+responses = sim.responses.load(
+    config,
+    model.RT.YTrans,
+    pier_settlement=[model.PierSettlement(0, 1.2)]
+)
+plot.contour_responses(config, responses)
+plot.top_view_bridge(config.bridge, piers=True)
+plt.tight_layout()
+plt.show()
 
-# Plotting different response types.
+################################################
+# Example 4: plotting different response types #
+################################################
 
-# import matplotlib.pyplot as plt
-# from bridge_sim import bridges, configs, model, plot, sim
+import matplotlib.pyplot as plt
+from bridge_sim import bridges, configs, model, plot, sim
 
-# config = configs.opensees_default(bridges.bridge_wide)
[…]
