+++ bridge_sim/plot/__init__.py
@@ -9,10 +9,10 @@ import matplotlib.pyplot as plt
 import numpy as np
 from matplotlib import patches as patches
 
+from bridge_sim.internal.plot import default_cmap, plt, axis_cmap_r
 from bridge_sim.model import Config, Vehicle, PointLoad, Point, Bridge
 from bridge_sim.sim.build import get_bridge_shells
 from bridge_sim.sim.model import Responses, SimParams, Shell
-from lib.plot import default_cmap, plt, axis_cmap_r
 
 import bridge_sim.plot.animate as animate
 
