+++ src/setup.py
@@ -1,3 +1,0 @@
-from distutils.core import setup
-
-setup(name="bridge_sim", version="0.1")
