+++ bridge_sim/internal/make/plot/distribution.py
@@ -1,429 +1,0 @@
-import itertools
-from typing import List
-
-import numpy as np
-
-from bridge_sim.internal.plot import plt
-from bridge_sim.internal.plot.responses import plot_distributions
-from bridge_sim.model import Config, Point, ResponseType
-from bridge_sim.traffic import normal_traffic
-from bridge_sim.sim.responses import to_traffic_array
-from bridge_sim.scenarios import HealthyScenario, Scenario
-from bridge_sim.sim.model import Responses
-from bridge_sim.sim.run.opensees import OSRunner
-from bridge_sim.plot import contour_responses
-from bridge_sim.util import print_i, safe_str
-
-
-def load_normal_traffic_array(c: Config, mins: float, lam: float, min_d: float):
-    """Distribution plots currently use this for normal traffic.
-
-    TODO: Move to classify.traffic.
-
-    """
-    traffic_scenario = normal_traffic(c, lam=lam, min_d=min_d)
-    return load_traffic_array(
-        c=c, traffic_scenario=traffic_scenario, max_time=60 * mins
-    )
-
-
-def distribution_plots(c: Config):
-    """Make all distribution plots."""
-    lane_distribution_plots(
-        c=c,
-        bridge_scenarios=(all_scenarios(c)),
-        response_type=ResponseType.YTranslation,
-    )
-    pier_displacement_distribution_plots(c=c, response_type=ResponseType.YTranslation)
-
-
-def lane_distribution_plots(
-    c: Config,
-    bridge_scenarios: List[Scenario],
-    response_type: ResponseType,
-    num: int = 25,
-):
-    """For each 'DamageScenario' plot response distributions along each lane.
-
-    All simulations are under the normal traffic scenario.
-
-    Args:
-        c: Config, global configuration object.
-        bridge_scenarios: List[DamageScenario], each bridge scenario for which
-            to plot the distribution of fem. The first scenario must be
-            the healthy scenario.
-        response_type: ResponseType, the type of sensor response to record.
-        num: int, the number of points at which to record fem.
-
-    """
-    assert isinstance(bridge_scenarios[0], HealthyScenario)
-    print_i("Lane distribution plots: loading traffic")
-    normal_traffic_array, traffic_scenario = load_normal_traffic_array(c)
-
-    def lane_points(lane):
-        """For points along half the lane."""
-        return [
-            Point(x=x, y=0, z=lane.z_center())
-            for x in np.linspace(c.bridge.x_min, c.bridge.x_max / 2, num)
-        ]
-
-    healthy_responses = []
-    response_arrays = []
-    amin, amax = np.inf, -np.inf
-
-    # Collect fem for each bridge scenario and lane combination.
-    for b, bridge_scenario in enumerate(bridge_scenarios):
-        print_i(f"Lane distribution plots: bridge scenario {bridge_scenario.name}")
-        for lane_index, lane in enumerate(c.bridge.lanes):
-            print_i(f"Lane distribution plots: lane {lane_index}")
-            points = lane_points(lane)
-            response_arrays.append(
-                to_traffic_array(
-                    config=c,
-                    traffic_array=normal_traffic_array,
-                    response_type=response_type,
-                    bridge_scenario=bridge_scenario,
-                    points=points,
-                    fem_runner=OSRunner(c),
-                )
-            )
-            if b == 0:
-                healthy_responses.append(response_arrays[-1])
-            resized, _ = resize_and_units(response_arrays[-1], response_type)
-            maybe_min, maybe_max = np.amin(resized), np.amax(resized)
-            if maybe_min < amin:
-                amin = maybe_min
-            if maybe_max > amax:
-                amax = maybe_max
-
-    assert len(c.bridge.lanes) == len(healthy_responses)
-
[…]
