+++ bridge_sim/sim/crack.py
@@ -4,11 +4,11 @@ from typing import List
 import bridge_sim.util
 import numpy as np
 
+from bridge_sim.internal.plot import plt
 from bridge_sim.model import Config, Point
 from bridge_sim.sim.responses import to_traffic_array
 from bridge_sim.traffic import load_traffic, normal_traffic
 from bridge_sim import temperature
-from lib.plot import plt
 from bridge_sim.util import safe_str
 
 
