### 2020-08-11 0bbda44
Include data files in PyPI package
