+++ CHANGELOG.MD
@@ -1,2 +1,7 @@
+# [WIP] 0.0.9
+- Stiffness parameters for piers.
+- Automatically publish Docker image.
+- Automatically publish on PyPI.
+- Package published on PyPI.
 # 0.0.8
-Moved from pipenv to poetry.
+- Moved from pipenv to poetry.
