+++ bridge_sim/internal/plot/validation.py
@@ -1,0 +1,170 @@
+import itertools
+from typing import Callable, Dict, Optional
+
+import matplotlib
+import numpy as np
+import pandas as pd
+
+from bridge_sim.model import Config, Point
+from bridge_sim.sim.model import Responses
+from bridge_sim.internal.plot import plt
+from bridge_sim.util import print_i, scalar
+
+
+def plot_mmm_strain_convergence(
+    c: Config,
+    pier: int,
+    df: pd.DataFrame,
+    all_strains: Dict[float, Responses],
+    title: str,
+    without: Optional[Callable[[Point], bool]] = None,
+    append: Optional[str] = None,
+):
+    """Plot convergence of given fem as model size grows."""
+    # A grid of points 1m apart, over which to calculate fem.
+    grid = [
+        Point(x=x, y=0, z=z)
+        for x, z in itertools.product(
+            np.linspace(c.bridge.x_min, c.bridge.x_max, int(c.bridge.length)),
+            np.linspace(c.bridge.z_min, c.bridge.z_max, int(c.bridge.width)),
+        )
+    ]
+    # If requested, remove some values from the fem.
+    if without is not None:
+        grid = [point for point in grid if not without(point)]
+        for msl, strains in all_strains.items():
+            print(f"Removing points from strains with max_shell_len = {msl}")
+            all_strains[msl] = strains.without(without)
+    # Collect fem over all fem, and over the grid. Iterate by
+    # decreasing max_shell_len.
+    mins, maxes, means = [], [], []
+    gmins, gmaxes, gmeans = [], [], []
+    max_shell_lens = []
+    for msl, strains in sorted(all_strains.items(), key=lambda kv: -kv[0]):
+        max_shell_lens.append(msl)
+        print_i(f"Gathering strains with max_shell_len = {msl}", end="\r")
+        grid_strains = np.array([strains.at_deck(point, interp=True) for point in grid])
+        gmins.append(scalar(np.min(grid_strains)))
+        gmaxes.append(scalar(np.max(grid_strains)))
+        gmeans.append(scalar(np.mean(grid_strains)))
+        strains = np.array(list(strains.values()))
+        mins.append(scalar(np.min(strains)))
+        maxes.append(scalar(np.max(strains)))
+        means.append(scalar(np.mean(strains)))
+    print()
+    # Normalize and plot the mins, maxes, and means.
+    def normalize(ys):
+        print(ys)
+        return ys / np.mean(ys[-5:])
+
+    mins, maxes, means = normalize(mins), normalize(maxes), normalize(means)
+    gmins, gmaxes, gmeans = normalize(gmins), normalize(gmaxes), normalize(gmeans)
+    deck_nodes = [df.at[msl, "deck-nodes"] for msl in max_shell_lens]
+    pier_nodes = [df.at[msl, "pier-nodes"] for msl in max_shell_lens]
+    num_nodes = np.array(deck_nodes) + np.array(pier_nodes)
+    print(f"MSLs = {max_shell_lens}")
+    print(f"num_nodes = {num_nodes}")
+    # Plot all lines, for debugging.
+    plt.landscape()
+    plt.plot(num_nodes, mins, label="mins")
+    plt.plot(num_nodes, maxes, label="maxes")
+    plt.plot(num_nodes, means, label="means")
+    plt.plot(num_nodes, gmins, label="gmins")
+    plt.plot(num_nodes, gmaxes, label="gmaxes")
+    plt.plot(num_nodes, gmeans, label="gmeans")
+    plt.grid(axis="y")
+    plt.xlabel("Nodes in FEM")
+    plt.ylabel("Strain")
+    plt.title(title)
+    plt.tight_layout()
+    plt.legend()
+    plt.savefig(
[…]
