+++ bridge_sim/internal/make/plot/verification.py
@@ -1,0 +1,1520 @@
+"""Verification plots."""
+
+import glob
+import itertools
+import os
+from collections import defaultdict
+from copy import deepcopy
+from datetime import datetime
+from timeit import default_timer as timer
+from typing import List, Optional, Tuple
+
+import matplotlib
+import matplotlib.cm as cm
+import numpy as np
+import pandas as pd
+from scipy.interpolate import interp1d
+from sklearn.linear_model import LinearRegression
+
+from bridge_sim.internal.plot import plt
+from bridge_sim.model import Config, PierSettlement, Point, PointLoad, ResponseType
+from bridge_sim.sim.responses import to_vehicles_direct
+from bridge_sim.vehicles import truck1
+from bridge_sim.sim.build import det_nodes, det_shells
+from bridge_sim.sim.model import SimParams
+from bridge_sim.sim.responses import load_fem_responses
+from bridge_sim.sim.run.opensees import OSRunner
+from bridge_sim.plot.util import legend_marker_size
+from bridge_sim.plot import contour_responses, top_view_bridge
+from bridge_sim.internal.plot.validation import (
+    plot_mmm_strain_convergence,
+    plot_nesw_convergence,
+)
+from bridge_sim.util import (
+    print_i,
+    print_w,
+    round_m,
+    safe_str,
+    scalar,
+)
+from bridge_sim.internal.validate import (
+    _meas,
+    _displa_sensors,
+    _strain_sensors,
+    _strain_sensor_xz,
+    _displa_sensor_xz,
+)
+
+# Positions of truck front axle.
+truck_front_x = np.arange(1, 116.1, 1)
+
+diana_path = "data/validation/modelpredictions_april2019.csv"
+if os.path.exists(diana_path):
+    diana = pd.read_csv(diana_path)
+
+# Interpolation function for each sensor position.
+diana_interp_funcs = dict()
+
+
+def diana_response(sensor_label: str, truck_x: float):
+    """Strain or displacement from Diana for a sensor and truck position."""
+    if sensor_label not in diana_interp_funcs:
+        all_sensors = diana[diana["sensorlabel"] == sensor_label]
+        diana_interp_funcs[sensor_label] = interp1d(
+            all_sensors["xpostruck"], all_sensors["infline1"]
+        )
+    return diana_interp_funcs[sensor_label](truck_x)
+
+
+def per_sensor_plots(
+    c: Config,
+    rows: int = 3,
+    strain_sensors_startwith: str = "T",
+    strain_sensors_ignore: List[str] = ["T0"],
+    individual_sensors: List[str] = ["T4", "U3"],
+    plot_diana: bool = False,
+):
+    """Compare the bridge 705 measurement campaign to Diana and OpenSees."""
+    plt.portrait()
+    size = 80  # Size of scatter plot points.
+    lw = 4  # Line width of plots.
+
+    ##########
+    # Strain #
+    ##########
+
+    print_i("All strain sensors = ")
+    print_i(f"  {sorted(set(_meas['sensorlabel']))}")
+    # All strain measurements for a given sensor set (strain_sensors_startwith),
+    # except ignore sensors in ignore set (strain_sensors_ignore).
+    tno_strain_meas = _meas.loc[
+        _meas["sensorlabel"].str.startswith(strain_sensors_startwith)
+    ]
+    labels_before_ignore = set(tno_strain_meas["sensorlabel"])
+    tno_strain_meas = tno_strain_meas.loc[
+        ~tno_strain_meas["sensorlabel"].isin(strain_sensors_ignore)
+    ]
+    labels_after_ignore = set(tno_strain_meas["sensorlabel"])
+    labels_ignored = sorted(
+        l for l in labels_before_ignore if l not in labels_after_ignore
+    )
[…]
