### 2020-08-11 7b411d4
Publish on pypi (#185)
* Publish on PyPI

* Update README.org: extend pypi installation guide

* Update docs

* Include data files in PyPI package

* Update tests for new data

Co-authored-by: Árpád Rózsás <rozsasarpi@gmail.com>
