+++ bridge_sim/internal/cli.py
@@ -1,0 +1,775 @@
+"""Command line interface to bridge-sim.
+
+Run './scripts/cli.sh' from the root directory of the cloned 'bridge_sim' repository.
+
+"""
+
+import os
+import pdb
+import sys
+import traceback
+
+import click
+
+from bridge_sim.configs import opensees_default
+from bridge_sim.internal.validate import _truck1_x_pos
+from bridge_sim.internal.make import paramselect
+from bridge_sim.internal.make.plot import contour as contour_
+from bridge_sim.internal.make.plot import geometry as geometry_
+from bridge_sim.internal.make.plot import vehicle, verification
+from bridge_sim.model import ResponseType
+from bridge_sim.vehicles import truck1, og_truck1
+from bridge_sim.bridges.bridge_705 import bridge_705
+from bridge_sim.util import clean_generated, print_i, remove_except_npy
+
+# Storing CLI parameters here as global variables.
+pdb_ = "--pdb" in sys.argv
+b_func = None
+c_func = opensees_default
+parallel_ = None
+save_to_ = None
+shorten_paths_ = None
+two_materials_ = None
+il_num_loads_ = None
+crack_x_ = None
+crack_len_ = None
+
+
+def c():
+    """Construct a "Config" based on CLI parameters."""
+    new_c = c_func(b_func)
+    if crack_x_ is not None:
+        print_i(f"Using transverse crack")
+        new_c = bridge_sim.crack.transverse_crack(
+            at_x=crack_x_, length=crack_len_
+        ).crack(new_c)
+    else:
+        print_i(f"Not using transverse crack")
+    new_c.parallel = parallel_
+    new_c.shorten_paths = shorten_paths_
+    new_c.il_num_loads = il_num_loads_
+    old_c_root_generated_data_dir = new_c._root_generated_data_dir
+    new_c.root_generated_data_dir = lambda: os.path.join(
+        save_to_, old_c_root_generated_data_dir
+    )
+    return new_c
+
+
+@click.group()
+@click.option(
+    "--uls", type=int, default=600, help="Unit load simulations per wheel track",
+)
+@click.option(
+    "--msl", type=float, default=0.5, help="Maximum shell length of the bridge.",
+)
+@click.option(
+    "--two-materials",
+    is_flag=True,
+    help="One material for the deck and one for piers.",
+)
+@click.option(
+    "--parallel", type=int, default=1, help="Run experiment simulations in parallel",
+)
+@click.option(
+    "--save-to", type=str, default="", help="Save/load data from a given folder.",
+)
+@click.option(
+    "--shorten-paths", is_flag=True, help="Save responses at shorter filepaths.",
+)
+@click.option(
+    "--pdb", is_flag=True, help="Jump into the debugger on exception.",
+)
+@click.option("--crack-x", type=float, help="Transverse crack at x.")
+@click.option("--crack-len", type=float, help="X length of transverse crack.")
+def cli(
+    uls: int,
+    msl: str,
+    two_materials: bool,
+    parallel: int,
+    save_to: bool,
+    shorten_paths: bool,
+    pdb: bool,
+    crack_x,
+    crack_len,
+):
+    global b_func
+    global save_to_
+    global parallel_
+    global shorten_paths_
+    global il_num_loads_
+    global two_materials_
+    global crack_x_
+    global crack_len_
+    crack_x_ = crack_x
+    crack_len_ = crack_len
+    two_materials_ = two_materials
+    b_func = bridge_705(msl=msl, single_sections=two_materials)
+    save_to_ = save_to
+    parallel_ = parallel
+    shorten_paths_ = shorten_paths
+    il_num_loads_ = uls
+
+    click.echo(f"Bridge: {b_func().name}")
[…]
