### 2020-08-11 4df40eb
Add more informative error messages when OS fails to run (#172)
* more informative error messages when OS cannot be run (particularly for Windows machines)
* Move PermissionError to non-FE specific code
