+++ .github/workflows/publish-pypi.yml
@@ -1,0 +1,14 @@
+name: Publish PyPI
+on:
+  push:
+    tags:
+      - 'v*.*.*'
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    - name: Publish
+      uses: JRubics/poetry-publish@v1.1
+      with:
+        pypi_token: ${{ secrets.PYPI_PW }}
