+++ .github/workflows/publish-pypi.yml
@@ -11,4 +11,4 @@ jobs:
     - name: Publish
       uses: JRubics/poetry-publish@v1.1
       with:
-        pypi_token: ${{ secrets.PYPI_PW }}
+        pypi_token: ${{ secrets.pypi_token }}
