### 2020-08-11 cff6495
Publish pypi action (#187)
* Action to publish to PyPI
