### 2020-08-14 8c0bb51
Add LineSpringSupport class
