### 2020-08-14 62acce5
Add duplicate nodes at bottom of piers
