### 2020-08-14 50999ad
Generate TCL with duplicate nodes
