+++ bridge_sim/sim/run/opensees/build/d3/__init__.py
@@ -15,6 +15,8 @@ from bridge_sim.sim.model import (
     Node,
     PierNodes,
     PierShells,
+    UniaxialMaterial,
+    ZeroLengthElement,
     SimParams,
 )
 from bridge_sim.sim.build import (
@@ -179,43 +181,123 @@ def opensees_fixed_abutment_nodes(
     )
 
 
-def opensees_uniaxial_material(material_id, stiffness):
+def opensees_uniaxial_material(
+    material_id,
+    stiffness,
+    damping_tangent: float = 0.0,
+    stiffness_under_compression: float = None,
+):
     """OpenSees UniaxialMaterial TCL command.
 
     https://opensees.berkeley.edu/wiki/index.php/Elastic_Uniaxial_Material
-
+    TODO: check units!
     """
     if not stiffness_under_compression:
         stiffness_under_compression = stiffness
-    return f"uniaxialMaterial Elastic {material_id} {stiffness * 1E6}"
+    return (
+        f"uniaxialMaterial Elastic {material_id} {stiffness * 1E6} "
+        f"{damping_tangent} {stiffness_under_compression}"
+    )
 
 
-def opensees_pier_boundary_conditions(c: Config, all_pier_nodes: PierNodes, ctx: BuildContext):
+def opensees_pier_boundary_conditions(
+    c: Config, all_pier_nodes: PierNodes, ctx: BuildContext
+):
     """OpenSees commands for stiffness properties at the bottom of each pier."""
-    dupe_bottom_nodes, uniaxial_materials, zero_length_elements = [], [], []
+    (
+        all_bottom_nodes,
+        all_dupe_bottom_nodes,
+        all_uniaxial_materials,
+        all_zero_length_elements,
+    ) = ([], [], [], [])
     # For the nodes of each pier.
     for p_i, pier_nodes in enumerate(all_pier_nodes):
-        # TODO: Add uniaxial materials.
-        # TODO: Add zerolength elements.
-        pier = c.bridge.supports[p_i]
-        # For each bottom node.
-        for y_i, y_nodes in enumerate(pier_nodes[0]):
-            bottom_node = y_nodes[-1]
-            dupe_bottom_nodes.append(ctx.get_node(
-                x=bottom_node.x,
-                y=bottom_node.y,
-                z=bottom_node.z,
-                deck=bottom_node.deck,
-                comment=bottom_node.comment,
-                nth_node=2,
-            ))
-    # TODO: Return uniaxial materials string.
-    # TODO: Return zerolength elements string.
-    return ""
-    return comment(
+        stiffness_z_rotation = c.bridge.supports[p_i].support.stiffness_z_rotation
+        if stiffness_z_rotation != 0:
+            pier = c.bridge.supports[p_i]
+            bottom_nodes, dupe_bottom_nodes, zero_length_elements = [], [], []
+            n_pier_bottom_node = len(pier_nodes[0])
+            bottom_node_coords = np.empty((n_pier_bottom_node, 3))
+            bottom_node_ids = np.empty(n_pier_bottom_node, dtype=int)
+            dupe_bottom_node_ids = np.empty(n_pier_bottom_node, dtype=int)
+            # For each bottom node.
+            for y_i, y_nodes in enumerate(pier_nodes[0]):
+                bottom_node = y_nodes[-1]
+                dupe_bottom_node = ctx.get_node(
+                    x=bottom_node.x,
+                    y=bottom_node.y,
+                    z=bottom_node.z,
+                    deck=bottom_node.deck,
+                    comment=bottom_node.comment,
+                    nth_node=2,
+                )
+                bottom_nodes.append(bottom_node)
+                dupe_bottom_nodes.append(dupe_bottom_node)
[…]
