### 2020-08-16 8e7a620
rename a string ('dir') that was modified during refactoring
