### 2020-08-16 6286a1e
take only the length along the z direction
