### 2020-08-17 72bdaad
Avoid duplicate UniaxialMaterials
