### 2020-08-17 1c7f951
Fix get_node's handling of nth_node arg
