### 2020-08-19 1d93280
Update Ubuntu-based Dockerfile
