### 2020-08-19 5d3982d
Update Ubuntu-based Dockerfile
