### 2020-08-19 cd95406
Testing new Docker image
