### 2020-08-19 f866847
Remove uneeded file
