+++ scripts/docker.sh
@@ -1,6 +1,6 @@
 CAT << EOF
 docker login --username=barischrooneyj
-docker build -f docker/Dockerfile
+docker build -f docker/Dockerfile .
 docker tag Y barischrooneyj/bridge-sim:version
 docker push barischrooneyj/bridge-sim:version
 EOF
