+++ docker/Dockerfile
@@ -1,8 +1,8 @@
 FROM ubuntu:20.04
-ENV OPENSEES_TAG v3.2.0
 ENV BRIDGE_SIM_TAG v0.0.8.6
+ENV OPENSEES_TAG v3.2.0
+ENV DEBIAN_FRONTEND noninteractive
 ENV PATH /root/.poetry/bin:/home/bin:$PATH
-ENV DEBIAN_FRONTEND=noninteractive
 # Install OpenSees.
 RUN apt-get update && \
     apt-get install -y --no-install-recommends ca-certificates g++ gcc gfortran git make \
@@ -15,9 +15,9 @@ RUN apt-get update && \
     cp MAKES/Makefile.def.EC2-UBUNTU Makefile.def && \
     sed -i 's#BASE\t\t= ./usr/local#BASE\t\t= /usr/local#' Makefile.def && \
     sed -i 's#HOME\t\t= ./home#HOME\t\t= /home#' Makefile.def && \
-    make
-# Install bridge-sim.
-RUN cd /home && \
+    make && \
+    # Install bridge-sim.
+    cd /home && \
     apt-get install -y --no-install-recommends curl ffmpeg python3 python3-venv && \
     curl -sSL https://raw.githubusercontent.com/python-poetry/poetry/master/get-poetry.py \
       | python3 && \
@@ -26,9 +26,10 @@ RUN cd /home && \
     git checkout $BRIDGE_SIM_TAG && \
     poetry install --no-dev && \
     poetry run python3 scripts/dockerize-example.py && \
-    poetry run python3 example.py
-# Cleanup.
-RUN cd /home && \
+    poetry run python3 example.py && \
+    # Cleanup.
+    cd /home && \
     rm -r OpenSees lib && \
-    apt-get remove ca-certificates curl git make python3-dev tcl8.6-dev && \
+    apt-get remove -y ca-certificates curl git make python3-dev tcl8.6-dev && \
+    rm -rf /var/lib/apt/lists/*
 WORKDIR /home/bridge-sim
