### 2020-08-19 6545207
Test auto-publish to Docker hub
