+++ .github/workflows/publish-pypi.yml
@@ -1,14 +1,30 @@
-name: Publish PyPI
+name: Publish on PyPI and Docker Hub
 on:
   push:
     tags:
       - 'v*.*.*'
 jobs:
-  build:
+  publish_on_pypi:
+    name: Publish on PyPI
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v2
-    - name: Publish
+    - name: Checkout code
+      uses: actions/checkout@v2
+    - name: Build and publish
       uses: JRubics/poetry-publish@v1.1
       with:
         pypi_token: ${{ secrets.PYPI_TOKEN }}
+  publish_on_docker_hub:
+    name: Publish on Docker Hub
+    needs: publish_on_pypi
+    runs-on: ubuntu-latest
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v2
+    - name: Build and push Docker images
+      uses: docker/build-push-action@v1
+      with:
+        username: ${{ secrets.DOCKER_USERNAME }}
+        password: ${{ secrets.DOCKER_PASSWORD }}
+        repository: barischrooneyj/bridge-sim
+        tag_with_ref: true
