### 2020-08-19 2aeffd2
Test auto-publish to Docker hub
