### 2020-08-19 70b6954
Test auto-publish to Docker hub
