### 2020-08-21 0eb2123
Update bridge_sim/sim/run/opensees/build/d3/__init__.py
Co-authored-by: Árpád Rózsás <rozsasarpi@gmail.com>
