### 2020-08-21 0b91b5a
Add fixed duplicate node
