+++ pyproject.toml
@@ -26,6 +26,7 @@ numpy = "1.19.0rc2"
 pandas = "1.0.4"
 sklearn = "0.0"
 termcolor = "1.1.0"
+orderedset = "^2.0.3"
 
 [tool.poetry.dev-dependencies]
 black = "19.10b0"
