### 2020-08-27 #2496 @andybond13
Learning how to use docker...

FYI: As instructed in the documentation, "docker run -it barischrooneyj/bridge-sim:v0.0.8.5" returns error "docker: Error response from daemon: manifest for barischrooneyj/bridge-sim:v0.0.8.5 not found: manifest unknown: manifest unknown."

using "docker run -it barischrooneyj/bridge-sim:v0.0.8" seems to work.
