### 2020-09-02 e43b617
Test examples in README
