### 2020-09-02 c492c1d
Test examples in README
