### 2020-09-02 998f895
Update bridge-sim version in docker iamge
