### 2020-09-13 #2496 @andybond13
Hello @diehlpk , I have completed my review. I apologize for the delayed completion!

I am not comfortable accepting the submission at this point, primarily due to the lack of broad applicability. The submission implicitly contains many limiting assumptions which would prevent adoption of the submitted code outside of the author’s group, let alone to professional bridge engineers. Hence, my recommendation would be to suggest that the author make significant changes before publication.

I have prepared a detailed review (attached), which maps comments to the review checklist above.

[JOSS_review_sept13_2020.pdf](https://github.com/openjournals/joss-reviews/files/5215219/JOSS_review_sept13_2020.pdf)
