### 2020-09-13 #2496 @barischrooneyj
@andybond13 Thank you for a very informative review! I will work through it, it might take 6-12 months since A I am not working on this project full time and B additional functionality will be added to bridge-sim. However it will get done. Thanks again :smile:
