### 2020-09-14 #2496 @diehlpk
Hi, @andybond13 thanks for the intense review. I agree that the paper/code is not ready for submission yet.

I will pause the review for now and once @barischrooneyj finished the major revision, I will open it again. @barischrooneyj just ping me here once you have the revision done. In addition, please respond to all of the reviewer comments.

 I am looking forward to your revision.
