### 2020-10-22 d334ccd
Informative comments to node positioning
