+++ scripts/test.sh
@@ -1,3 +1,4 @@
+#!/usr/bin/env bash
 set -euo pipefail
 
 rm -rf test-data/
