+++ scripts/cli.sh
@@ -1,1 +1,4 @@
+#!/usr/bin/env bash
+set -euo pipefail
+
 poetry run python bridge_sim/internal/cli.py "$@"
