+++ pyproject.toml
@@ -2,9 +2,9 @@
 name = "bridge-sim"
 version = "0.0.8.8"
 description = "A Python library for concrete slab bridge simulation"
-authors = ["Jeremy Barisch-Rooney <barischrooneyj@protonmail.com>"]
-documentation = "https://github.com/barischrooneyj/bridge-sim/blob/master/README.org#concrete-slab-bridge-simulation"
-repository = "https://github.com/barischrooneyj/bridge-sim"
+authors = ["Jeremy Barisch-Rooney <jerbaroo.work@pm.me>"]
+documentation = "https://github.com/jerbaroo/bridge-sim/blob/master/README.org#concrete-slab-bridge-simulation"
+repository = "https://github.com/jerbaroo/bridge-sim"
 include = [
     "model-template-3d.tcl",
     "data/temperature/holly-springs-18.txt",
