### 2021-01-08 991e8b6
Fix version regex in release script
