+++ scripts/release.sh
@@ -2,12 +2,12 @@
 set -euo pipefail
 
 # Check versions to migrate to/from are correct.
-RE='^[0-9]+([.][0-9]+)?$'
+RE='^[0-9]+([.][0-9]+)*$'
 if ! [[ "$1" =~ $RE ]] ; then
-  echo "error: first argument not a number" >&2; exit 1
+  echo "error: first argument is not a number" >&2; exit 1
 fi
 if ! [[ "$2" =~ $RE ]] ; then
-  echo "error: second argument not a number" >&2; exit 1
+  echo "error: second argument is not a number" >&2; exit 1
 fi
 read -r -p "Migrating from v$1 to v$2, correct? [y/N]" RESPONSE 
 case "$RESPONSE" in
@@ -42,10 +42,12 @@ export BRIDGE_SIM_TESTS=STANDARD
 ./scripts/test.sh
 
 # Replace version numbers in files.
-sed -i '' -e "s/v$1/v$2/g" README.org docker/Dockerfile pyproject.toml
+sed -i '' -e "s/$1/$2/g" README.org docker/Dockerfile pyproject.toml
 echo "Replaced v$1 with v$2"
 
 # Final instructions.
+echo 'Check "git diff" manually'
+git diff
 echo 'Finally run these commands:'
 echo "git commit -am 'Bump version to v$2'"
-echo "git push"
+echo 'git push'
